mcp-server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.13.0
pydantic-settings==2.4.0
pydantic_core==2.46.0
python-dotenv==1.2.2
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0


fastapi==0.115.0          # Web framework for building APIs
uvicorn==0.30.6           # ASGI server that runs FastAPI apps
httpx==0.27.2             # Async HTTP client for external API calls
python-dotenv==1.0.1      # Reads .env files into environment variables
pydantic==2.9.2           # Data validation (used by FastAPI internally)